Lock World Scribe 2 Server dependency to specific tag
Previously, the world-scribe-2-server dependency was tied to the master
branch. This is bad for build integrity: if a build of the desktop app
is re-run with the same code at 2 different points in time, the builds
might have wildly different functionality because the server's master
branch updated sometime in between builds.

To get around this, we explicitly set a git tag to pull from. To get
new changes from the server repo, this tag reference must be
explicitly updated in package.json. This makes it easy to track
changes in Git.

In the future, it would be ideal to automate this process so that
whenever a new tag is pushed to the server repo, the desktop repo's
package.json is updated to reference the new tag. We could likely
use Github Actions to accomplish this.